+++ util/snowflake.go
@@ -38,7 +38,7 @@ func NewSnowflake(workerID int64) (*Snowflake, error) {
 }
 
 // NextId 下一个ID
-func (w *Snowflake) NextId() uint64 {
+func (w *Snowflake) NextId() uint {
 	w.mu.Lock()
 	defer w.mu.Unlock()
 	now := time.Now().UnixNano() / 1e6
@@ -53,7 +53,7 @@ func (w *Snowflake) NextId() uint64 {
 		w.number = 0
 		w.timestamp = now
 	}
-	Id := uint64((now-startTime)<<timeShift | (w.workerID << workerShift) | (w.number))
+	Id := uint((now-startTime)<<timeShift | (w.workerID << workerShift) | (w.number))
 	return Id
 }
 
@@ -65,6 +65,6 @@ func init() {
 	_snowflake, _ = NewSnowflake(0xFF)
 }
 
-func PrimaryKey() uint64 {
+func PrimaryKey() uint {
 	return _snowflake.NextId()
 }
